feat!: add TerminalBackend.resizeFull() for receiving terminal size in pixels
In addition to terminal width and height in the unit of
columns and rows provided by TerminalBackend.resize(),
now there is a TerminalBackend.resizeFull() method that
can be implemented to receive terminal width and height
in pixels in additional to in columns and rows. This is
to fulfill the winsize struct in different Unix platforms.

In order to ensure minimum API changes,
TerminalBackend.resizeFull() is implemented in a way
that existing TerminalBackend implementations does not
need to resizeFull() if they do not need the additional
information. Instead they can just implement resize()
to fulfill the TerminalBackend interface.

BREAKING CHANGE: Implementations of TerminalBackend should
    now extends TerminalBackend instead of implements
    TerminalBackend. Simply changing implements to extends
    will be enough to resolve the breaking API changes.