Generalizes dirty handling on a BufferLine
Now anyone can define a Tag (String) to mark a line as "clean". The clean / dirty state can be retrieved by also providing that tag.
The Bufferline will remove that clean marker whenever the content of that line changes.
This way we can get rid of the Function handlers that set and query each dirty state separately from a line and replace that by just expecting a unique tag name.